Wave 2: Fix hook bugs and convert CJS files to ESM
- Fix useCurrencyExchangeRates: catch (erorr) typo made error undefined
- Fix useSubscriptions: rename getSubscritions->getSubscriptions, add try/catch
  so UI doesn't hang on auth/fetch failures
- Convert firebase/admin.js, firebase/index.js to ESM imports/exports
- Convert scripts/firebase/firestoreSeed.js -> .mjs with ESM imports;
  fix forEach(async) anti-pattern to sequential for-of loop
