+++ firebase/index.js
@@ -1,6 +1,6 @@
-const { initializeApp } = require("firebase/app");
-const { getFirestore } = require("firebase/firestore");
-const { getAuth } = require("firebase/auth");
+import { initializeApp } from "firebase/app";
+import { getFirestore } from "firebase/firestore";
+import { getAuth } from "firebase/auth";
 
 const firebaseConfig = {
   apiKey: process.env.NEXT_PUBLIC_API_KEY,
@@ -13,7 +13,5 @@ const firebaseConfig = {
 
 const app = initializeApp(firebaseConfig);
 
-module.exports = {
-  db: getFirestore(app),
-  auth: getAuth(app),
-};
+export const db = getFirestore(app);
+export const auth = getAuth(app);
